Fix int type of is date function

+++ cpp/src/gandiva/gdv_function_stubs.cc
@@ -183,7 +183,7 @@ CAST_NUMERIC_FROM_STRING(double, arrow::DoubleType, FLOAT8)
 
 #undef CAST_NUMERIC_FROM_STRING
 
-bool gdv_fn_is_date_utf8(int64_t ptr, const char* data, int data_len) {
+bool gdv_fn_is_date_utf8(int64_t ptr, const char* data, int64_t data_len) {
   return ::arrow::internal::ParseTimestampStrptime(
       data, data_len, "%Y-%m-%d", /*ignore_time_in_day=*/false,
       /*allow_trailing_chars=*/false, ::arrow::TimeUnit::SECOND, &ptr);
